Reduce scope of serialization and deserialization
