+++ GPTifier/src/api.cpp
@@ -1,8 +1,8 @@
 #include "api.hpp"
+
 #include "configs.hpp"
 
 #include <cstdlib>
-#include <curl/curl.h>
 #include <fmt/core.h>
 #include <json.hpp>
 #include <optional>
@@ -45,34 +45,7 @@ void catch_curl_error(CURLcode code)
     }
 }
 
-class Curl {
-public:
-    Curl();
-    ~Curl();
-
-    // We want to prevent any copies from being made otherwise we'll attempt
-    // to delete a shallow copy of the headers list multiple times (i.e. because the destructor will
-    // be called for each copy)
-    Curl(const Curl &) = delete;
-    Curl &operator=(const Curl &) = delete;
-
-    std::string get_uploaded_files();
-    std::string get_models();
-    std::string create_chat_completion(const std::string &post_fields);
-    std::string create_embedding(const std::string &post_fields);
-    std::string upload_file(const std::string &filename, const std::string &purpose);
-    std::string delete_file(const std::string &file_id);
-    std::string create_fine_tuning_job(const std::string &post_fields);
-    std::string delete_model(const std::string &model_id);
-    std::string get_fine_tuning_jobs(const std::string &limit);
-
-    CURL *handle = NULL;
-
-private:
-    void set_project_id();
-
-    struct curl_slist *headers = NULL;
-};
+} // namespace
 
 Curl::Curl()
 {
@@ -285,8 +258,6 @@ std::string Curl::get_fine_tuning_jobs(const std::string &limit)
     return response;
 }
 
-} // namespace
-
 namespace api {
 
 std::string create_chat_completion(const std::string &model, const std::string &prompt, float temperature)
@@ -298,14 +269,6 @@ std::string create_chat_completion(const std::string &model, const std::string &
 
     Curl curl;
     return curl.create_chat_completion(data.dump());
-}
-
-std::string create_embedding(const std::string &model, const std::string &input)
-{
-    const nlohmann::json data = { { "model", model }, { "input", input } };
-
-    Curl curl;
-    return curl.create_embedding(data.dump());
 }
 
 std::string get_uploaded_files()
